refactor(wardens): extract glob + command-parse leaf capsules (LIA-306)
Begin the incremental decomposition of the 4261-line live gate engine
scripts/codex_warden_hooks.py into a scripts/warden_hooks/ package. First
slice extracts two pure-stdlib leaf clusters behind re-exports, so the
runtime path and the public/private symbol surface stay identical:

- warden_hooks/globs.py:         _glob_to_regex, _glob_match
- warden_hooks/command_parse.py: _command_hash, _shell_tokens,
  _gh_command_index_after_global_flags, _is_gh_executable,
  _is_admin_merge_command, _extract_pr_ref

The entry module re-imports them and drops two now-orphaned stdlib
imports (functools, PureWindowsPath). No behavior change.

Verified: 380 warden tests pass on Python 3.12 (CI-parity venv) and 3.14;
hook-pr-smoke-test shows byte-identical real-invocation output
(admin-merge-gate incl. command hash, structural-check) vs pre-refactor.

LIA-306 step 1 of N; remaining capsules deferred to sequential follow-ups.
